Moved mkdocs_macros code
